third_party/blink/web_tests/crypto/subtle/aes-kw/cloneKey-expected.txt

Tests structured cloning of AES-KW keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-KW"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-KW"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "wrapKey,unwrapKey"
Serialized key bytes: 5c4b010c10c1011030112233445566778899aabbccddeeff
PASS: Cloned key exported data should be [30112233445566778899aabbccddeeff] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-KW"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-KW"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "wrapKey,unwrapKey"
Serialized key bytes: 5c4b010c20c1012000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f
PASS: Cloned key exported data should be [00112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-KW"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-KW"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "wrapKey,unwrapKey"
Serialized key bytes: 5c4b010c10c0011030112233445566778899aabbccddeeff


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-KW"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-KW"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "wrapKey,unwrapKey"
Serialized key bytes: 5c4b010c20c0012000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f


PASS successfullyParsed is true

TEST COMPLETE
